Add files via upload
Used for decrypting en mass when encrypted with the pdf_paranoia.py script. Instead of making an entirely new program as listed in the book I tried to reuse as much code from pdf_paranoia.py as possible. This took a bit of re-writing and general cleaning up but as far as I can tell it does what the spec laid out.